+++ src/core/src/main/java/com/qlkh/core/client/model/TaskQuota.java
@@ -15,41 +15,41 @@ import com.qlkh.core.client.model.core.AbstractEntity;
 public class TaskQuota extends AbstractEntity {
 
     private Task task;
-    private double q1;
-    private double q2;
-    private double q3;
-    private double q4;
+    private int q1;
+    private int q2;
+    private int q3;
+    private int q4;
     private int year;
 
-    public double getQ1() {
+    public int getQ1() {
         return q1;
     }
 
-    public void setQ1(double q1) {
+    public void setQ1(int q1) {
         this.q1 = q1;
     }
 
-    public double getQ2() {
+    public int getQ2() {
         return q2;
     }
 
-    public void setQ2(double q2) {
+    public void setQ2(int q2) {
         this.q2 = q2;
     }
 
-    public double getQ3() {
+    public int getQ3() {
         return q3;
     }
 
-    public void setQ3(double q3) {
+    public void setQ3(int q3) {
         this.q3 = q3;
     }
 
-    public double getQ4() {
+    public int getQ4() {
         return q4;
     }
 
-    public void setQ4(double q4) {
+    public void setQ4(int q4) {
         this.q4 = q4;
     }
